move packages and fix code to work with new openapi generated code
