+++ src/main/kotlin/org/developerden/codosseum/server/Routes.kt
@@ -2,7 +2,7 @@ package org.developerden.codosseum.server
 
 import io.github.tabilzad.ktor.annotations.KtorFieldDescription
 import io.ktor.resources.*
-import org.developerden.codosseum.indexing.Challenge
+import org.developerden.codosseum.challenges.indexing.Challenge
 
 @Resource("/validate")
 class Validate {
